XX-6489 configure personal assistant IM account
Personal assistant's IM account is automatically created by sipXconfig.
The account information can be configured using sipXivr service
configuration page. Default name of the assistant "MyAssistant" and
default password is empty.

Personal assistant's account information is propagated to
sipxivr.properties and xmpp-acounts.xml files.

Based on patch from: Constantin Tambrea - thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16794 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXconfig/plugins/openfire/src/org/sipfoundry/sipxconfig/openfire/SipxOpenfireConfiguration.java
@@ -80,7 +80,7 @@ public class SipxOpenfireConfiguration extends SipxServiceConfiguration {
             String port = StringUtils.substringAfterLast(server, HOST_PORT_SEPERATOR);
             if (port.isEmpty() || Integer.valueOf(port) < SERVER_TO_SERVER_PORT_MIN
                     || Integer.valueOf(port) > SERVER_TO_SERVER_PORT_MAX) {
-                port = String.valueOf((Integer) getSipxOpenfireService().getSettingTypedValue(
+                port = String.valueOf(getSipxOpenfireService().getSettingTypedValue(
                         SipxOpenfireService.SERVER_TO_SERVER_DEFAULT_REMOTE_PORT));
             }
